refactor: migrate stitch cli to dataio.open_files_and_stitch function
